Fix auto fill and write tests
